+++ CHANGELOG.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `pctx add` now accepts `--header` and `--bearer` to add authentication without interaction
+- `pctx.json` config now accepts version which gets returned as the MCP's version in the `initialize` MCP response
 - add typescript type check runtime capabilities including more typical string/array utils
 - tool descriptions updated for consistent behavior
-
 
 ### Fixed